Computed Properties

+++ app/Livewire/ArticleList.php
@@ -3,7 +3,10 @@
 namespace App\Livewire;
 
 use App\Models\Article;
+use Exception;
+use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\View\View;
+use Livewire\Attributes\Computed;
 use Livewire\Attributes\Title;
 use Livewire\WithPagination;
 
@@ -16,7 +19,15 @@ class ArticleList extends AdminComponent
 
     public function delete(Article $article)
     {
+        if ($this->articles->count() < 10) {
+            throw new Exception('Nope');
+        }
+
         $article->delete();
+
+        unset($this->articles);
+
+        cache()->forget('published-count');
     }
 
     public function showAll(): void
@@ -33,14 +44,18 @@ class ArticleList extends AdminComponent
 
     public function render(): View
     {
+        return view('livewire.article-list');
+    }
+
+    #[Computed(persist: true)]
+    protected function articles(): LengthAwarePaginator
+    {
         $query = Article::query();
 
         if ($this->showOnlyPublished) {
             $query->where('published', 1);
         }
 
-        return view('livewire.article-list', [
-            'articles' => $query->paginate(10, pageName: 'articles-page'),
-        ]);
+        return  $query->paginate(10, pageName: 'articles-page');
     }
 }
